create arch module
