+++ arch/build.gradle
@@ -33,18 +33,13 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation Dependencies.androidX_app_compat
-    implementation Dependencies.androidX_localbroadcast_manager
     implementation Dependencies.androidX_core_ktx
-    implementation Dependencies.gson
+    implementation Dependencies.androidX_recyclerView
 
     //lifecycle
     implementation Dependencies.androidX_lifecycle_viewModel
     implementation Dependencies.androidX_lifecycle_extensions
     kapt Dependencies.androidX_lifecycle_compiler
-
-    //testing
-    androidTestImplementation Dependencies.androidX_test_runner
-    testImplementation Dependencies.jUnit
 
     //dagger2
     implementation Dependencies.dagger_core
@@ -53,7 +48,13 @@ dependencies {
     kapt Dependencies.dagger_compiler
     kapt Dependencies.dagger_android_processor
 
+    //testing
+    androidTestImplementation Dependencies.androidX_test_runner
+    testImplementation Dependencies.jUnit
+
     //kotlin
     implementation Dependencies.kotlin_reflect
     implementation Dependencies.kotlin_stdlib
+
+    implementation project(":core")
 }
