+++ arch/src/main/java/knaufdan/android/arch/dagger/ArchModule.kt
@@ -1,0 +1,15 @@
+package knaufdan.android.arch.dagger
+
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class ArchModule {
+
+    @Singleton
+    @Provides
+    internal fun provideNavigator(navigationService: knaufdan.android.arch.navigation.NavigationService): knaufdan.android.arch.navigation.INavigationService =
+        navigationService
+
+}
