+++ arch/src/main/java/knaufdan/android/arch/databinding/Translator.kt
@@ -1,0 +1,21 @@
+package knaufdan.android.arch.databinding
+
+import androidx.databinding.InverseMethod
+
+object Translator {
+
+    @JvmStatic
+    @InverseMethod("fromIntToString")
+    fun fromStringToInt(value: String): Int? {
+        return try {
+            Integer.parseInt(value)
+        } catch (e: NumberFormatException) {
+            null
+        }
+    }
+
+    @JvmStatic
+    fun fromIntToString(value: Int?): String {
+        return value?.toString() ?: ""
+    }
+}
