+++ arch/src/main/java/knaufdan/android/arch/mvvm/IActivityViewModel.kt
@@ -1,0 +1,16 @@
+package knaufdan.android.arch.mvvm
+
+import androidx.fragment.app.FragmentManager
+import knaufdan.android.arch.mvvm.implementation.BaseActivity
+
+typealias isHandled = Boolean
+
+interface IActivityViewModel : IBaseViewModel {
+
+    /**
+     * Is called on [BaseActivity.onBackPressed].
+     *
+     * @return [true] if the backPress event is fully handled and must not be forwarded via [super.onBackPressed].
+     */
+    fun handleBackPressed(fragmentManager: FragmentManager): isHandled
+}
