+++ arch/src/main/java/knaufdan/android/arch/mvvm/IAndroidComponent.kt
@@ -1,0 +1,5 @@
+package knaufdan.android.arch.mvvm
+
+interface IAndroidComponent {
+    fun getTitleRes(): Int = -1
+}
