+++ arch/src/main/java/knaufdan/android/arch/mvvm/IBaseFragment.kt
@@ -1,0 +1,9 @@
+package knaufdan.android.arch.mvvm
+
+import knaufdan.android.arch.mvvm.implementation.BaseViewModel
+import knaufdan.android.arch.databinding.BindableElement
+
+interface IBaseFragment<ViewModel : BaseViewModel> : BindableElement<ViewModel>,
+    IAndroidComponent {
+    fun setBackPressed(isBackPressed: Boolean)
+}
