+++ arch/src/main/java/knaufdan/android/arch/mvvm/implementation/Config.kt
@@ -1,0 +1,32 @@
+package knaufdan.android.arch.mvvm.implementation
+
+import androidx.annotation.LayoutRes
+import androidx.annotation.StringRes
+import knaufdan.android.arch.navigation.FragmentContainer
+
+internal sealed class Config(
+    @LayoutRes val layoutRes: Int,
+    val viewModelKey: Int,
+    @StringRes val titleRes: Int
+) {
+    internal class ActivityConfig(
+        layoutRes: Int,
+        viewModelKey: Int,
+        titleRes: Int,
+        val fragmentSetup: Pair<FragmentContainer, BaseFragment<out BaseViewModel>?>?
+    ) : Config(
+        layoutRes,
+        viewModelKey,
+        titleRes
+    )
+
+    internal class FragmentConfig(
+        layoutRes: Int,
+        viewModelKey: Int,
+        titleRes: Int
+    ) : Config(
+        layoutRes,
+        viewModelKey,
+        titleRes
+    )
+}
