+++ arch/src/main/java/knaufdan/android/arch/navigation/INavigationService.kt
@@ -1,0 +1,23 @@
+package knaufdan.android.arch.navigation
+
+typealias FragmentContainer = Int
+
+interface INavigationService {
+    var fragmentContainer: FragmentContainer
+
+    /**
+     * Clears backStack and sets [fragment] into the [container].
+     */
+    fun cleanGoTo(
+        fragment: knaufdan.android.arch.mvvm.implementation.BaseFragment<out knaufdan.android.arch.mvvm.implementation.BaseViewModel>,
+        container: FragmentContainer = fragmentContainer
+    )
+
+    fun goTo(
+        fragment: knaufdan.android.arch.mvvm.implementation.BaseFragment<out knaufdan.android.arch.mvvm.implementation.BaseViewModel>,
+        addToBackStack: Boolean,
+        container: FragmentContainer = fragmentContainer
+    )
+
+    fun onBackPressed()
+}
