+++ arch/src/main/java/knaufdan/android/arch/navigation/NavigationService.kt
@@ -1,0 +1,61 @@
+package knaufdan.android.arch.navigation
+
+import android.app.Activity
+import androidx.appcompat.app.AppCompatActivity
+import javax.inject.Inject
+import javax.inject.Singleton
+import knaufdan.android.core.ContextProvider
+
+@Singleton
+class NavigationService @Inject constructor(private val contextProvider: ContextProvider) :
+    INavigationService {
+    override var fragmentContainer = -1
+
+    override fun cleanGoTo(
+        fragment: knaufdan.android.arch.mvvm.implementation.BaseFragment<out knaufdan.android.arch.mvvm.implementation.BaseViewModel>,
+        container: FragmentContainer
+    ) = with(contextProvider.context) {
+
+        check(container != -1) { "Could not replace ${fragment.fragmentTag} because no fragmentContainer is defined. Current container value = $container" }
+
+        if (this is AppCompatActivity) {
+            supportFragmentManager.popBackStackImmediate()
+            goTo(
+                fragment = fragment,
+                addToBackStack = false,
+                container = container
+            )
+        }
+    }
+
+    override fun goTo(
+        fragment: knaufdan.android.arch.mvvm.implementation.BaseFragment<out knaufdan.android.arch.mvvm.implementation.BaseViewModel>,
+        addToBackStack: Boolean,
+        container: FragmentContainer
+    ) = with(contextProvider.context) {
+
+        check(container != -1) { "Could not replace ${fragment.fragmentTag} because no fragmentContainer is defined. Current container value = $container" }
+
+        if (this is AppCompatActivity) {
+            supportFragmentManager.beginTransaction().apply {
+                replace(
+                    container,
+                    fragment,
+                    fragment.fragmentTag
+                )
+
+                if (addToBackStack) {
+                    addToBackStack(null)
+                }
+
+                commitAllowingStateLoss()
+            }
+        }
+    }
+
+    override fun onBackPressed() = with(contextProvider.context) {
+        if (this is Activity) {
+            onBackPressed()
+        }
+    }
+}
