--- core/src/main/java/knaufdan/android/core/arch/IAndroidComponent.kt
@@ -1,5 +1,0 @@
-package knaufdan.android.core.arch
-
-interface IAndroidComponent {
-    fun getTitleRes(): Int = -1
-}
