--- core/src/main/java/knaufdan/android/core/arch/IBaseFragment.kt
@@ -1,8 +1,0 @@
-package knaufdan.android.core.arch
-
-import knaufdan.android.core.arch.implementation.BaseViewModel
-import knaufdan.android.core.databinding.BindableElement
-
-interface IBaseFragment<ViewModel : BaseViewModel> : BindableElement<ViewModel>, IAndroidComponent {
-    fun setBackPressed(isBackPressed: Boolean)
-}
