--- core/src/main/java/knaufdan/android/core/arch/IBaseViewModel.kt
@@ -1,13 +1,0 @@
-package knaufdan.android.core.arch
-
-import android.os.Bundle
-import androidx.lifecycle.LifecycleObserver
-import knaufdan.android.core.arch.implementation.BaseFragment
-
-interface IBaseViewModel : LifecycleObserver {
-
-    /**
-     * Is called in [BaseActivity.onCreate] and [BaseFragment.onCreateView] before binding the [IBaseViewModel].
-     */
-    fun handleBundle(bundle: Bundle?)
-}
