--- core/src/main/java/knaufdan/android/core/arch/implementation/ActivityViewModel.kt
@@ -1,9 +1,0 @@
-package knaufdan.android.core.arch.implementation
-
-import androidx.fragment.app.FragmentManager
-import knaufdan.android.core.arch.IActivityViewModel
-
-abstract class ActivityViewModel : BaseViewModel(), IActivityViewModel {
-
-    override fun handleBackPressed(fragmentManager: FragmentManager) = false
-}
