--- core/src/main/java/knaufdan/android/core/databinding/BindableViewHolder.kt
@@ -1,15 +1,0 @@
-package knaufdan.android.core.databinding
-
-import androidx.databinding.ViewDataBinding
-import androidx.recyclerview.widget.RecyclerView
-
-class BindableViewHolder<DataSource>(
-    private val binding: ViewDataBinding,
-    private val bindingKey: BindingKey
-) : RecyclerView.ViewHolder(binding.root) {
-
-    fun bind(dataSource: DataSource) {
-        binding.setVariable(bindingKey, dataSource)
-        binding.executePendingBindings()
-    }
-}
