--- core/src/main/java/knaufdan/android/core/databinding/Translator.kt
@@ -1,21 +1,0 @@
-package knaufdan.android.core.databinding
-
-import androidx.databinding.InverseMethod
-
-object Translator {
-
-    @JvmStatic
-    @InverseMethod("fromIntToString")
-    fun fromStringToInt(value: String): Int? {
-        return try {
-            Integer.parseInt(value)
-        } catch (e: NumberFormatException) {
-            null
-        }
-    }
-
-    @JvmStatic
-    fun fromIntToString(value: Int?): String {
-        return value?.toString() ?: ""
-    }
-}
