+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':core'
+include ':core', ':arch'
 rootProject.name='ArchServices'
